Added weighted heuristic.  Made heuristic values in [0,1] by convention

+++ src/main/java/org/ggp/dhtp/util/Heuristic.java
@@ -6,7 +6,7 @@ import org.ggp.base.util.statemachine.exceptions.MoveDefinitionException;
 import org.ggp.base.util.statemachine.exceptions.TransitionDefinitionException;
 
 public abstract class Heuristic {
-	public abstract int evalState(Role role, MachineState state) throws TransitionDefinitionException, MoveDefinitionException, GoalDefinitionException;
+	public abstract double evalState(Role role, MachineState state) throws TransitionDefinitionException, MoveDefinitionException, GoalDefinitionException;
 }
